% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dadosmrpotec.R
\docType{data}
\name{dadosmrpotec}
\alias{dadosmrpotec}
\title{Documentação da base dadosmr}
\format{
description
}
\usage{
dadosmrpotec
}
\description{
A short description...
}
\author{
Felipe Smolski
}
\keyword{datasets}
